requirements/test.txt
nose
django-nose
nose-progressive
coverage
httmock
